pyproject.toml
[project]
name = "wallwatch"
version = "0.1.0"
description = "Order book wall monitor for T-Invest gRPC"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "WallWatch"}]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
]

dependencies = [
  "t-tech-investments>=0.3.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "mypy>=1.7",
]
telegram = [
  "python-telegram-bot>=20.7",
]

[project.scripts]
wallwatch = "wallwatch.app.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"
